+++ flask_sqlalchemy_bundle/validation.py
@@ -1,3 +1,4 @@
+from flask_unchained import _
 from flask_unchained.string_utils import title_case
 from typing import *
 
@@ -63,4 +64,6 @@ class Required(BaseValidator):
         return True
 
     def get_message(self, e: ValidationError):
+        if self.msg and isinstance(self.msg, str):
+            return self.msg
         return f'{title_case(e.column)} is required'